+++ tests/lib/ocpp/common/CMakeLists.txt
@@ -3,4 +3,5 @@ target_sources(libocpp_unit_tests PRIVATE
     test_database_migration_files.cpp
     test_database_schema_updater.cpp
     test_message_queue.cpp
+    utils_tests.cpp
 )
