+++ tests/lib/ocpp/common/test_message_queue.cpp
@@ -4,11 +4,6 @@
 #include <gtest/gtest.h>
 #include <nlohmann/json.hpp>
 #include <ocpp/common/message_queue.hpp>
-#include <ocpp/v16/messages/Authorize.hpp>
-#include <ocpp/v16/messages/MeterValues.hpp>
-#include <ocpp/v16/messages/SecurityEventNotification.hpp>
-#include <ocpp/v16/messages/StartTransaction.hpp>
-#include <ocpp/v201/messages/Authorize.hpp>
 
 namespace ocpp {
 
@@ -134,89 +129,6 @@ template <> bool ControlMessage<TestMessageType>::is_transaction_update_message(
 
 bool is_boot_notification_message(const TestMessageType message_type) {
     return message_type == TestMessageType::BootNotification;
-}
-
-/************************************************************************************************
- * ControlMessage
- *
- * Test implementations of ControlMessage template
- */
-class ControlMessageV16Test : public ::testing::Test {
-
-protected:
-};
-
-TEST_F(ControlMessageV16Test, test_is_transactional) {
-
-    EXPECT_TRUE(is_transaction_message(
-        ControlMessage<v16::MessageType>{Call<v16::StartTransactionRequest>{v16::StartTransactionRequest{}, "0"}}
-            .messageType));
-    EXPECT_TRUE(is_transaction_message(
-        ControlMessage<v16::MessageType>{Call<v16::StopTransactionRequest>{v16::StopTransactionRequest{}, "0"}}
-            .messageType));
-    EXPECT_TRUE(is_transaction_message(ControlMessage<v16::MessageType>{
-        Call<v16::SecurityEventNotificationRequest>{v16::SecurityEventNotificationRequest{}, "0"}}
-                                           .messageType));
-    EXPECT_TRUE(is_transaction_message(
-        ControlMessage<v16::MessageType>{Call<v16::MeterValuesRequest>{v16::MeterValuesRequest{}, "0"}}.messageType));
-
-    EXPECT_TRUE(!is_transaction_message(
-        ControlMessage<v16::MessageType>{Call<v16::AuthorizeRequest>{v16::AuthorizeRequest{}, "0"}}.messageType));
-}
-
-TEST_F(ControlMessageV16Test, test_is_transactional_update) {
-
-    EXPECT_TRUE(
-        !(ControlMessage<v16::MessageType>{Call<v16::StartTransactionRequest>{v16::StartTransactionRequest{}, "0"}})
-             .is_transaction_update_message());
-    EXPECT_TRUE(
-        !(ControlMessage<v16::MessageType>{Call<v16::StopTransactionRequest>{v16::StopTransactionRequest{}, "0"}})
-             .is_transaction_update_message());
-    EXPECT_TRUE(!(ControlMessage<v16::MessageType>{
-                      Call<v16::SecurityEventNotificationRequest>{v16::SecurityEventNotificationRequest{}, "0"}})
-                     .is_transaction_update_message());
-    EXPECT_TRUE((ControlMessage<v16::MessageType>{Call<v16::MeterValuesRequest>{v16::MeterValuesRequest{}, "0"}})
-                    .is_transaction_update_message());
-
-    EXPECT_TRUE(!(ControlMessage<v16::MessageType>{Call<v16::AuthorizeRequest>{v16::AuthorizeRequest{}, "0"}})
-                     .is_transaction_update_message());
-}
-
-class ControlMessageV201Test : public ::testing::Test {
-
-protected:
-};
-
-TEST_F(ControlMessageV201Test, test_is_transactional) {
-
-    EXPECT_TRUE(is_transaction_message(
-        ControlMessage<v201::MessageType>{Call<v201::TransactionEventRequest>{v201::TransactionEventRequest{}, "0"}}
-            .messageType));
-
-    EXPECT_TRUE(!is_transaction_message(
-        ControlMessage<v201::MessageType>{Call<v201::AuthorizeRequest>{v201::AuthorizeRequest{}, "0"}}.messageType));
-}
-
-TEST_F(ControlMessageV201Test, test_is_transactional_update) {
-
-    v201::TransactionEventRequest transaction_event_request{};
-    transaction_event_request.eventType = v201::TransactionEventEnum::Updated;
-
-    EXPECT_TRUE((ControlMessage<v201::MessageType>{Call<v201::TransactionEventRequest>{transaction_event_request, "0"}})
-                    .is_transaction_update_message());
-
-    transaction_event_request.eventType = v201::TransactionEventEnum::Started;
-    EXPECT_TRUE(
-        !(ControlMessage<v201::MessageType>{Call<v201::TransactionEventRequest>{transaction_event_request, "0"}})
-             .is_transaction_update_message());
-
-    transaction_event_request.eventType = v201::TransactionEventEnum::Ended;
-    EXPECT_TRUE(
-        !(ControlMessage<v201::MessageType>{Call<v201::TransactionEventRequest>{transaction_event_request, "0"}})
-             .is_transaction_update_message());
-
-    EXPECT_TRUE(!(ControlMessage<v201::MessageType>{Call<v201::AuthorizeRequest>{v201::AuthorizeRequest{}, "0"}})
-                     .is_transaction_update_message());
 }
 
 /************************************************************************************************
