+++ tests/lib/ocpp/v16/CMakeLists.txt
@@ -1,3 +1,7 @@
+target_include_directories(libocpp_unit_tests PUBLIC 
+        ${CMAKE_CURRENT_SOURCE_DIR}
+)
+
 target_sources(libocpp_unit_tests PRIVATE
         profile_tests_common.cpp
         profile_testsA.cpp
@@ -5,4 +9,6 @@ target_sources(libocpp_unit_tests PRIVATE
         profile_testsC.cpp
         test_database_migration_files.cpp
         test_smart_charging_handler.cpp
-)
+        database_tests.cpp
+        test_message_queue.cpp
+        )
