+++ tests/lib/ocpp/v201/CMakeLists.txt
@@ -1,4 +1,6 @@
-target_include_directories(libocpp_unit_tests PUBLIC mocks)
+target_include_directories(libocpp_unit_tests PUBLIC 
+        mocks
+        ${CMAKE_CURRENT_SOURCE_DIR})
 
 target_sources(libocpp_unit_tests PRIVATE
         test_database_migration_files.cpp
@@ -7,4 +9,7 @@ target_sources(libocpp_unit_tests PRIVATE
         test_ocsp_updater.cpp
         test_component_state_manager.cpp
         test_device_model.cpp
-        test_smart_charging_handler.cpp)
+        test_smart_charging_handler.cpp
+        utils_tests.cpp
+        comparators.cpp
+        test_message_queue.cpp)
